🐛 [Fix] hover 이벤트 위치 수정
